backend_django/she_aware/apps/education/serializers.py
from rest_framework import serializers
from .models import Article

class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = ['id', 'title', 'slug', 'category', 'body_html', 'sources', 'tags', 'lang', 'updated_at']
