backend_django/she_aware/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
google-generativeai>=0.3
python-dotenv>=1.0
requests
